Refactor behavior out of the app namespace
This behavior will be used for apps and plugins now, so make it more general.
