+++ pakyow-core/lib/pakyow/app.rb
@@ -11,18 +11,18 @@ require "pakyow/support/hookable"
 require "pakyow/support/configurable"
 require "pakyow/support/pipelined"
 
-require "pakyow/app/behavior/config"
-require "pakyow/app/behavior/cookies"
-require "pakyow/app/behavior/sessions"
-require "pakyow/app/behavior/endpoints"
-require "pakyow/app/behavior/pipeline"
-require "pakyow/app/behavior/frameworks"
-require "pakyow/app/behavior/aspects"
-require "pakyow/app/behavior/helpers"
-require "pakyow/app/behavior/rescuing"
-require "pakyow/app/behavior/restarting"
-require "pakyow/app/behavior/isolating"
-require "pakyow/app/behavior/initializers"
+require "pakyow/behavior/config"
+require "pakyow/behavior/cookies"
+require "pakyow/behavior/sessions"
+require "pakyow/behavior/endpoints"
+require "pakyow/behavior/pipeline"
+require "pakyow/behavior/frameworks"
+require "pakyow/behavior/aspects"
+require "pakyow/behavior/helpers"
+require "pakyow/behavior/rescuing"
+require "pakyow/behavior/restarting"
+require "pakyow/behavior/isolating"
+require "pakyow/behavior/initializers"
 
 require "pakyow/connection"
