+++ pakyow-routing/lib/pakyow/behavior/security/helpers.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+require "pakyow/support/extension"
+
+require "pakyow/security/helpers/csrf"
+
+module Pakyow
+  module Security
+    module Behavior
+      module Helpers
+        extend Support::Extension
+
+        apply_extension do
+          register_helper :passive, Security::Helpers::CSRF
+        end
+      end
+    end
+  end
+end
