+++ pakyow-routing/lib/pakyow/behavior/security/insecure.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require "pakyow/support/extension"
+
+require "pakyow/security/errors"
+
+module Pakyow
+  module Security
+    module Behavior
+      module Insecure
+        extend Support::Extension
+
+        apply_extension do
+          handle InsecureRequest, as: 403 do
+            trigger(403)
+          end
+        end
+      end
+    end
+  end
+end
